tests/submission/PhilipTaylor/canvas/2d.shadow.pattern.alpha.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.shadow.pattern.alpha</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.shadow.html">shadow</a>.<a href="index.2d.shadow.pattern.html">pattern</a>.alpha</h1>
<p class="desc">Shadows are drawn correctly for partially-transparent fill patterns</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.shadow.render">2d.shadow.render</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="2d.shadow.pattern.alpha.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Shadows are drawn correctly for partially-transparent fill patterns");
_addTest(function(canvas, ctx) {

var pattern = ctx.createPattern(document.getElementById('transparent50.png'), 'repeat');
ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.shadowOffsetY = 50;
ctx.shadowColor = '#00f';
ctx.fillStyle = pattern;
ctx.fillRect(0, -50, 100, 50);

_assertPixelApprox(canvas, 50,25, 127,0,127,255, "50,25", "127,0,127,255", 2);


});
</script>
<img class="resource" id="transparent50.png" src="../images/transparent50.png">

<div id="log"></div>
</body></html>
